requirements.txt
django
djangorestframework
dj-rest-auth
django-allauth
djangorestframework-simplejwt
gunicorn
whitenoise
pymysql
cryptography